requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
spacy>=3.6.0
nltk>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
